+++ BoostPlugin.php
@@ -25,7 +25,7 @@ class BoostPlugin extends BasePlugin
 
     public function getVersion()
     {
-        return '0.3.x';
+        return '2.x';
     }
 
     public function getDeveloper()
@@ -45,12 +45,33 @@ class BoostPlugin extends BasePlugin
             'devDbName' => array(AttributeType::String, 'default' => ''),
             'stageDbName' => array(AttributeType::String, 'default' => ''),
             'dbName' => array(AttributeType::String, 'default' => ''),
+            'devDbUser' => array(AttributeType::String, 'default' => ''),
+            'stageDbUser' => array(AttributeType::String, 'default' => ''),
+            'dbUser' => array(AttributeType::String, 'default' => ''),
+            'devDbPass' => array(AttributeType::String, 'default' => ''),
+            'stageDbPass' => array(AttributeType::String, 'default' => ''),
+            'dbPass' => array(AttributeType::String, 'default' => ''),
+            'devDbHost' => array(AttributeType::String, 'default' => ''),
+            'stageDbHost' => array(AttributeType::String, 'default' => ''),
+            'dbHost' => array(AttributeType::String, 'default' => 'localhost'),
             'envRoot' => array(AttributeType::String, 'default' => ''),
             'vcsCache' => array(AttributeType::String, 'default' => ''),
             'vcsDirs' => array(AttributeType::String, 'default' => ''),
             'vcsUrl' => array(AttributeType::String, 'default' => ''),
             'preDeploymentHooks' => array(AttributeType::String, 'default' => ''),
             'postDeploymentHooks' => array(AttributeType::String, 'default' => ''),
+
+            // For Imarc, should be www-data:web
+            'resetOwnership' => array(AttributeType::String, 'default' => 'www-data:web'),
+
+            // For Imarc, should be g+rw
+            'resetPermissions' => array(AttributeType::String, 'default' => 'g+rw'),
+
+            // For Imarc, should be false
+            'keepDatabase' => array(AttributeType::Bool, 'default' => false),
+
+            // For Imarc, should be web@localhost
+            'resetDbPermissions' => array(AttributeType::String, 'default' => 'web@localhost'),
         );
     }